fonder refactor + main page loading
