+++ frontend/src/components/HomeView/HolidayPreview.tsx
