+++ frontend/src/components/NewHolidayView/PollBuilder.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import PollOption from "@/components/PollOption";
+import PollOption from "@/components/NewHolidayView/PollOption";
 import { Poll } from "@/types";
 
 type PollBuilderProps = {
