+++ frontend/src/components/NewHolidayView/PollOption.tsx
